requirements.txt
# Core Flask
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.1.0
phonenumbers==8.13.22

# Rate Limiting & Protection
Flask-Limiter==3.5.0
Flask-CORS==4.0.0

# Utilities
python-dotenv==1.0.0
PyJWT==2.8.0
redis==5.0.1

# Email
Flask-Mail==0.9.1

# Development
black==23.11.0
flake8==6.1.0